+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 -   Added .todo, .log, and .logs to the --build ignore list
 -   Added --skip_lint argument to gulp lint tasks
 -   Updated docker general-cli to v1.1.0
+-   Updated composer packages
 
 ### Added